Fix tests for session management on notification sync
* The test which was ensuring the fact that session is saved after decryption
was wrong, it didn't fail even when code to save session was removed. This is
tackled by try to decrypt the same message again.

* The test which was ensuring the fact that session is only saved when saving
messages succeeds was broken as it was using a session stored in memory. To
force this from not happened, we now re-open the cryptobox. The test could have
closed the session, but that wouldn't be correct as the CLI must crash on any
failure. Catching these failures is probably too much work for now.