Use haskellPackages.shellFor to get dev shell
This commit also removes a lot of cabal code which was also pinning and
constraining packages. All the pinning and constraining is now done in nixpkgs.
Doing this ensures that cabal doesn't try to build anything and just trusts the
ghc-pkg provided by nix to be the packages it needs.

Unfortunately, this drops support for being able to run builds on a non-nix
system. If it becomes a problem, maybe a script can be written to generate a
cabal.project and a cabal.project.freeze to ensure builds run. But that is a
problem to be solved when it becomes one.